Backend/requirements.txt
fastapi==0.115.11
uvicorn==0.35.0
tensorflow==2.19.0
pillow==11.1.0
python-multipart==0.0.20
numpy==1.26.4